clang-compatibility

+++ src/runner_doiact_grav.h
@@ -582,7 +582,7 @@ void runner_dosub_grav(struct runner *r, struct cell *ci, struct cell *cj,
     /* Get the opening angle theta. */
     float dx[3], theta;
     for (k = 0; k < 3; k++) {
-      dx[k] = fabsf(ci->loc[k] - cj->loc[k]);
+      dx[k] = fabs(ci->loc[k] - cj->loc[k]);
       if (periodic && dx[k] > 0.5 * s->dim[k]) dx[k] = -dx[k] + s->dim[k];
       if (dx[k] > 0.0f) dx[k] -= ci->h[k];
     }
